+++ src/omv/img/point.c
@@ -9,7 +9,7 @@
 #include "imlib.h"
 #include "xalloc.h"
 
-point_t *point_alloc(int x, int y)
+point_t *point_alloc(int16_t x, int16_t y)
 {
     point_t *p = xalloc(sizeof(point_t));
     p->x = x;
